requirements_lin.txt
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
ExifRead==3.0.0
filelock==3.13.1
fire==0.5.0
fsspec==2023.10.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
pandas==1.5.3
Pillow==10.1.0
pyarmor==8.4.7
pyarmor.cli.core==5.4.3
pybboxes==0.1.6
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sahi==0.11.15
shapely==2.0.2
six==1.16.0
sympy==1.12
termcolor==2.3.0
terminaltables==3.1.10
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.0.7
